+++ hutool-extra/src/test/java/cn/hutool/extra/ssh/JschUtilTest.java
@@ -29,7 +29,7 @@ public class JschUtilTest {
 	public void sftpTest() {
 		Session session = JschUtil.getSession("looly.centos", 22, "root", "123456");
 		Sftp sftp = JschUtil.createSftp(session);
-		sftp.mkDirs("/etc/test/aaa/bbb");
+		sftp.mkDirs("/opt/test/aaa/bbb");
 		Console.log("OK");
 	}
 }